requirements.txt
Django>=4.2
djangorestframework>=3.14
drf-nested-routers>=0.93.5
drf-spectacular>=0.27.2
drf-spectacular-sidecar>=2024.7.1
django-cors-headers>=4.4.0
Pillow>=10.4.0

celery>=5.4.0
redis>=5.0.8
boto3>=1.35.0

numpy>=1.26.0
trimesh>=4.4.9
open3d>=0.18.0
